Update Westlaw.js, add annotation support
Adds support for annotations: retrieves all highlights from case, fetches page number (using one-star "preferred pagination," so be careful if you are using a different reporter) on which each highlight begins, and stores a note on the Juris-M item of the form 
"HIGHLIGHTED TEXT1 [*NUMBER]"
"HIGHLIGHTED TEXT2 [*NUMBER]"
etc.

Fixes some really crappy code I wrote when I first did this: Jurisdiction parsing no longer breaks when dealing with federal district courts with multiple divisions. (Before, "N.D. Georgia, Southern District" or similar would cause the translator to treat the court as a state court.)

Plays better with WL citations (and NCBC citations, which use the same form). Before, "2020 WL 102020" would be treated just like a normal reporter citation. Now, it is correctly treated as an archive entry. This makes a difference for short-form citations. The short form of reporter citations is VOLUME REPORTER at PINCITE, but archives are not continuously paginated, so they follow the form ARCHIVENUM ARCHIVE RECORDNUM at PINCITE.

Generates a good "short title" from the full title. To do this, it uses the first party name (unless that party is a government or is much longer than the second party's name) and trims the useless crap (for example, "inc." and "LLP") from the end of the party's name. Note that the default Juris-M modules ignore the short title field for cases, so you will need to edit the relevant module to incorporate the short title.

Relies on the abbreviations filter (specifically its ability to conditionally suppress fields) rather than omitting metadata. In the past, cases whose court was unambiguously identified by the reporter name were imported without jurisdiction and court information, which was a hack to accomplish bluebook-correct formatting. I have since learned that this can be done using the Abbreviations filter.